+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-19 13:36:51
-// Commit: fix(core): fix memory leak by cancelling core subscription
+// Updated at: 2026-06-19 17:21:32
+// Commit: style(core): harmonize light and dark theme palette for core
 
 class coreComponentState {
-  final String commitRef = '2126';
-  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
-  final DateTime timestamp = DateTime.parse('2026-06-19T13:36:51+05:00');
+  final String commitRef = '2141';
+  final String lastMessage = 'style(core): harmonize light and dark theme palette for core';
+  final DateTime timestamp = DateTime.parse('2026-06-19T17:21:32+05:00');
 }